Create README
This repository hosts the open-source programming for Soliston, the next generation of IoT-enabled windows. While the patented latch mechanism is not part of the open-source initiative, the software here is designed to make the windows smart, customizable, and integrable with various smart home platforms. We aim to collaborate with the community to develop innovative solutions and expand functionality.